requirements.txt
PyYAML
Click
Jinja2
python-string-utils
gti-spe-lib-json-model